requirements.txt
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.6.2.post1
certifi==2024.8.30
charset-normalizer==3.4.0
contourpy==1.3.0
cycler==0.12.1
distro==1.9.0
exceptiongroup==1.2.2
filelock==3.16.1
fonttools==4.54.1
fsspec==2024.10.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
hydra-core==1.3.2
idna==3.10
iopath==0.1.10
Jinja2==3.1.4
jiter==0.6.1
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.7
langchain-core==0.3.13
langchain-openai==0.2.3
langsmith==0.1.137
MarkupSafe==3.0.2
matplotlib==3.9.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.1.2
omegaconf==2.3.0
openai==1.52.2
opencv-python==4.10.0.84
orjson==3.10.10
packaging==24.1
pillow==11.0.0
portalocker==2.10.1
psutil==6.1.0
pydantic==2.9.2
pydantic_core==2.23.4
pyparsing==3.2.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
requests-toolbelt==1.0.0
six==1.16.0
sniffio==1.3.1
sympy==1.13.1
tenacity==9.0.0
tiktoken==0.8.0
torch==2.5.0
torchvision==0.20.0
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.3
